x86: make fpu switching eager
Linux 4.14 dropped support for lazy fpu switching and in the 4.4 and 4.9
series similar changes where backported.
So we can assume that fpu switching is purely eager for those Linux
kernel versions. That simplifies things a lot and we can drop several
changes from the IPIPE patch.
On the Xenomai side the only thing we still have to care about is the
kernel fpu state when we interrupt an in kernel fpu user. But we do that
explicit and can drop the indirection active_(fp)state.

Signed-off-by: Henning Schild <henning.schild@siemens.com>